Added TranslationSource.specific_content_type field
TranslatableObjects can be create from a few different places and not
all of them will have access to the specific content type (for pages, we
can call page.specific, but other models that use
multi-table-inheritance might be used and there isn't a generic way to
get the specific type for those).

So to prevent any conflict, we always set
TranslatableObject.content_type to the content type which has the
`locale` and `translation_key` field. So we need a separate content type
field to allow TranslationSource to store the specific type. I think we
can safely assume that TranslationSource.from_instance() is always
passed the most specific content type.